Abstracted out damaging a whole view
