+++ compositor/xdg.c
@@ -52,54 +52,36 @@ static void wc_xdg_surface_commit(struct wl_listener* listener, void* data) {
 		return;
 	}
 
-	struct wlr_output* outputs[4] = { 0 };
-	wc_view_get_outputs(view->server->output_layout, view, outputs);
-
-	for (int i = 0; i < 4; i++) {
-		struct wlr_output* output = outputs[i];
-		if (output) {
-			struct wlr_box surface_area = {
-				.x = view->x - output->lx,
-				.y = view->y - output->ly,
-				.width = view->width,
-				.height = view->height
-			};
-			wlr_log(WLR_DEBUG, "DAMAGING (%d, %d : %d, %d)",
-					surface_area.x, surface_area.y, surface_area.width, surface_area.height);
-			struct wc_output* wc_output = output->data;
-			wlr_output_damage_add_box(wc_output->damage, &surface_area);
-		}
-	}
+	// TODO Only damage what's changed from the surface,
+	// then move this down into the pending if block below
+	wc_view_damage_whole(view);
 
 	struct wlr_xdg_surface* surface = view->xdg_surface;
+	struct wlr_box size = {0};
+	wlr_xdg_surface_get_geometry(surface, &size);
+
 	uint32_t pending_serial =
 		view->pending_serial;
-	if (pending_serial != 0 && pending_serial >+ surface->configure_serial) {
-		if (view->pending_geometry.x != view->x) {
-			view->x = view->pending_geometry.x;
-		}
-		if (view->pending_geometry.y != view->y) {
-			view->y = view->pending_geometry.y;
-		}
+	if (pending_serial != 0 && pending_serial >= surface->configure_serial) {
 		if (view->pending_geometry.width != view->width) {
 			view->width = view->pending_geometry.width;
 		}
 		if (view->pending_geometry.height != view->height) {
 			view->height = view->pending_geometry.height;
 		}
+		wc_view_damage_whole(view);
+		if (view->pending_geometry.x != view->x) {
+			view->x = view->pending_geometry.x +
+				view->pending_geometry.width - size.width;
+		}
+		if (view->pending_geometry.y != view->y) {
+			view->y = view->pending_geometry.y +
+				view->pending_geometry.height - size.height;
+		}
 		view->pending_serial = 0;
 	}
 
-	wc_view_get_outputs(view->server->output_layout, view, outputs);
-	// TODO Damage only what has changed
-	for (int i = 0; i < 4; i++) {
-		struct wlr_output* output = outputs[i];
-		if (output) {
-			wc_output_damage_surface(
-					output->data, view->xdg_surface->surface,
-					view->x - output->lx, view->y - output->ly);
-		}
-	}
+	wc_view_damage_whole(view);
 }
 
 static void wc_xdg_surface_destroy(struct wl_listener* listener, void* data) {
